+++ releases/v2/bitwrench-debug.js
@@ -1,4 +1,4 @@
-/*! bitwrench-debug v2.1.1 | BSD-2-Clause | https://deftio.github.io/bitwrench/pages */
+/*! bitwrench-debug v2.1.2 | BSD-2-Clause | https://deftio.github.io/bitwrench/pages */
 (function (global, factory) {
 	typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory() :
 	typeof define === 'function' && define.amd ? define(factory) :